'chessboard' program was created
'chessboard' program was created. This program  builds a chessboard with specified height and width.
After the building a chessboard user gets a message "Do you want to continue? [y/yes]"
The program will end its work or will continue depending on the user's response.
In the end user gets a message "The program is completed." It means that the program completed its work.