Added new modules to services
Added `converting_number_to_word` module to services. This file contains the logic for solving and running the task about converting number to word.
Added `lucky_tickets` module to services. This file contains the logic for solving and running the task about lucky tickets.
Added `triangles` module to services. This file contains the logic for solving and running the task about triangles sorting.